+++ changelog.md
@@ -12,6 +12,8 @@ Features
     + Fix the error of angle updating.
 + Show the values on expression table.
 + Add solution selection mode with expression table.
++ Dimensional synthesis function:
+    + Re-designed user interface.
 
 Development
 ---